festival, faq 연관관계 수정 #113
