+++ src/test/java/com/tsp/new_tsp_admin/api/faq/service/AdminFaqJpaServiceTest.java
@@ -371,7 +371,7 @@ class AdminFaqJpaServiceTest {
 
         AdminFaqDTO adminFaqDTO = AdminFaqEntity.toDto(adminFaqEntity);
 
-        adminFaqJpaService.updateFaq(adminFaqEntity);
+        adminFaqJpaService.updateFaq(idx, adminFaqEntity);
 
         // when
         when(mockAdminFaqJpaService.findOneFaq(adminFaqEntity.getIdx())).thenReturn(adminFaqDTO);
@@ -405,7 +405,7 @@ class AdminFaqJpaServiceTest {
 
         AdminFaqDTO adminFaqDTO = AdminFaqEntity.toDto(adminFaqEntity);
 
-        adminFaqJpaService.updateFaq(adminFaqEntity);
+        adminFaqJpaService.updateFaq(idx, adminFaqEntity);
 
         // when
         when(mockAdminFaqJpaService.findOneFaq(adminFaqEntity.getIdx())).thenReturn(adminFaqDTO);
