+++ src/test/java/com/tsp/new_tsp_admin/api/festival/service/AdminFestivalJpaQueryRepositoryTest.java
@@ -1,0 +1,111 @@
+package com.tsp.new_tsp_admin.api.festival.service;
+
+import com.tsp.new_tsp_admin.api.domain.festival.AdminFestivalDTO;
+import com.tsp.new_tsp_admin.api.domain.festival.AdminFestivalEntity;
+import lombok.RequiredArgsConstructor;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InOrder;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.event.EventListener;
+import org.springframework.test.context.TestConstructor;
+import org.springframework.test.context.TestPropertySource;
+
+import javax.persistence.EntityManager;
+import javax.transaction.Transactional;
+
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.*;
+import static org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace.NONE;
+import static org.springframework.test.context.TestConstructor.AutowireMode.ALL;
+
+@DataJpaTest
+@Transactional
+@TestPropertySource(locations = "classpath:application.properties")
+@TestConstructor(autowireMode = ALL)
+@RequiredArgsConstructor
+@AutoConfigureTestDatabase(replace = NONE)
+@ExtendWith(MockitoExtension.class)
+@DisplayName("모델 Repository Test")
+class AdminFestivalJpaQueryRepositoryTest {
+
+    @Mock
+    private AdminFestivalJpaQueryRepository mockAdminFestivalJpaQueryRepository;
+    private final AdminFestivalJpaQueryRepository adminFestivalJpaQueryRepository;
+    private final EntityManager em;
+
+    private AdminFestivalEntity adminFestivalEntity;
+    private AdminFestivalDTO adminFestivalDTO;
+
+    void createFestival() {
+        // 등록
+        LocalDateTime dateTime = LocalDateTime.now();
+
+        adminFestivalEntity = AdminFestivalEntity.builder()
+                .festivalTitle("축제 제목")
+                .festivalDescription("축제 내용")
+                .festivalMonth(dateTime.getMonthValue())
+                .festivalDay(dateTime.getDayOfMonth())
+                .festivalTime(dateTime)
+                .build();
+
+        em.persist(adminFestivalEntity);
+
+        adminFestivalDTO = AdminFestivalEntity.toDto(adminFestivalEntity);
+    }
+
+    @BeforeEach
+    @EventListener(ApplicationReadyEvent.class)
+    public void init() {
+        createFestival();
+    }
+
+    @Test
+    @DisplayName("축제 리스트 조회 테스트")
+    void 축제리스트조회테스트() {
+        Map<String, Object> festivalMap = new HashMap<>();
+        festivalMap.put("jpaStartPage", 1);
+        festivalMap.put("size", 3);
+
+        assertThat(adminFestivalJpaQueryRepository.findFestivalList(festivalMap)).isNotEmpty();
+    }
+
+    @Test
+    @DisplayName("축제 리스트 조회 Mockito 테스트")
+    void 축제리스트조회Mockito테스트() {
+        Map<String, Object> festivalMap = new HashMap<>();
+        festivalMap.put("jpaStartPage", 1);
+        festivalMap.put("size", 3);
+
+        List<AdminFestivalDTO> festivalList = new ArrayList<>();
+        festivalList.add(adminFestivalDTO);
+
+        // when
+        when(mockAdminFestivalJpaQueryRepository.findFestivalList(festivalMap)).thenReturn(festivalList);
+        List<AdminFestivalDTO> findFestivalList = mockAdminFestivalJpaQueryRepository.findFestivalList(festivalMap);
+
+        // then
+        assertThat(findFestivalList.get(0).getFestivalTitle()).isEqualTo("축제 제목");
+        assertThat(findFestivalList.get(0).getFestivalDescription()).isEqualTo("축제 내용");
+
+        // verify
+        verify(mockAdminFestivalJpaQueryRepository, times(1)).findFestivalList(festivalMap);
+        verify(mockAdminFestivalJpaQueryRepository, atLeastOnce()).findFestivalList(festivalMap);
+        verifyNoMoreInteractions(mockAdminFestivalJpaQueryRepository);
+
+        InOrder inOrder = inOrder(mockAdminFestivalJpaQueryRepository);
+        inOrder.verify(mockAdminFestivalJpaQueryRepository).findFestivalList(festivalMap);
+    }
+}
